make rate a floating point parameter
Note: adding a constructor with double rate is a breaking change, because calling TokenBucket(1, 1) would result in a compiler error "call of overloaded 'TokenBucket(int, int)' is ambiguous"

+++ TokenBucket.h
@@ -30,7 +30,7 @@ public:
   TokenBucket() {}
 
   /* rate: average number of tokens per second */
-  TokenBucket(const uint64_t rate, const uint64_t burstSize) {
+  TokenBucket(const double rate, const uint64_t burstSize) {
     timePerToken_ = 1000000 / rate;
     timePerBurst_ = burstSize * timePerToken_;
   }
